LPS-194437 Adapt AddObjectField modal to new React structure
