+++ modules/apps/object/object-web/types/src/main/resources/META-INF/resources/js/components/ObjectField/Fields.d.ts
@@ -8,6 +8,7 @@
 import {IFDSTableProps} from '../../utils/fds';
 interface FieldsProps extends IFDSTableProps {
 	baseResourceURL: string;
+	objectFieldTypes: ObjectFieldType[];
 }
 export default function Fields({
 	apiURL,
@@ -17,6 +18,7 @@ export default function Fields({
 	id,
 	items,
 	objectDefinitionExternalReferenceCode,
+	objectFieldTypes,
 	style,
 	url,
 }: FieldsProps): JSX.Element;
